--- parser/src/main/java/parser/DTO/Book.java
@@ -1,78 +1,0 @@
-package parser.DTO;
-
-import java.util.Set;
-
-/**
- * Created by grzegorz_sledz on 25.08.16.
- */
-public class Book {
-
-    private String title;
-    private String author;
-    private String price;
-    private String percentageDiscount;
-    private String genre;
-    private String description;
-    private Set<String> tags;
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public String getAuthor() {
-        return author;
-    }
-
-    public void setAuthor(String author) {
-        this.author = author;
-    }
-
-    public String getPrice() {
-        return price;
-    }
-
-    public void setPrice(String price) {
-        this.price = price;
-    }
-
-    public String getPercentageDiscount() {
-        return percentageDiscount;
-    }
-
-    public void setPercentageDiscount(String percentageDiscount) {
-        this.percentageDiscount = percentageDiscount;
-    }
-
-    public String getGenre() {
-        return genre;
-    }
-
-    public void setGenre(String genre) {
-        this.genre = genre;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public Set<String> getTags() {
-        return tags;
-    }
-
-    public void setTags(Set<String> tags) {
-        this.tags = tags;
-    }
-
-    @Override
-    public String toString() {
-        return "Book [title=" + title + ", author=" + author + ", price=" + price + ", percentageDiscount=" + percentageDiscount + ", genre=" + genre + " tags "+tags+ ", description=" + description + "]";
-    }
-}
